Readding tbpore/tbpore.py
