+++ tbpore/__init__.py
@@ -14,4 +14,3 @@ else:
     import importlib_metadata as metadata
 
 __version__ = metadata.version("tbpore")
-from tbpore.constants import *
